Make ExternalNavigator module a simple jvm one
Will only hold the interface and allow the :app module to create its
instance in the MainActivity directly

also move openEmailApp into ExternalNavigator in this process
