+++ app/app/src/main/kotlin/com/hedvig/android/app/di/ApplicationModule.kt
@@ -73,8 +73,6 @@ import com.hedvig.android.language.di.languageModule
 import com.hedvig.android.market.di.marketManagerModule
 import com.hedvig.android.market.di.setMarketModule
 import com.hedvig.android.memberreminders.di.memberRemindersModule
-import com.hedvig.android.navigation.activity.ExternalNavigator
-import com.hedvig.android.navigation.activity.ExternalNavigatorImpl
 import com.hedvig.android.navigation.core.HedvigDeepLinkContainer
 import com.hedvig.android.navigation.core.di.deepLinkModule
 import com.hedvig.android.notification.badge.data.di.notificationBadgeModule
@@ -174,12 +172,6 @@ fun makeUserAgent(languageBCP47: String): String = buildString {
   append(")")
 }
 
-private val externalNavigatorModule = module {
-  single<ExternalNavigator> {
-    ExternalNavigatorImpl(buildConfigApplicationId = BuildConfig.APPLICATION_ID)
-  }
-}
-
 private val buildConstantsModule = module {
   single<HedvigBuildConstants> {
     val context = get<Context>()
@@ -272,7 +264,6 @@ private val workManagerModule = module {
 val applicationModule = module {
   includes(
     listOf(
-      externalNavigatorModule,
       adyenFeatureModule,
       apolloAuthListenersModule,
       appModule,
