+++ app/app/src/main/kotlin/com/hedvig/android/app/ui/HedvigAppUi.kt
@@ -43,7 +43,6 @@ internal fun HedvigAppUi(
   finishApp: () -> Unit,
   shouldShowRequestPermissionRationale: (String) -> Boolean,
   openUrl: (String) -> Unit,
-  onOpenEmailApp: () -> Unit,
   market: Market,
   imageLoader: ImageLoader,
   languageService: LanguageService,
@@ -68,7 +67,6 @@ internal fun HedvigAppUi(
         finishApp = finishApp,
         shouldShowRequestPermissionRationale = shouldShowRequestPermissionRationale,
         openUrl = openUrl,
-        onOpenEmailApp = onOpenEmailApp,
         imageLoader = imageLoader,
         market = market,
         languageService = languageService,
