+++ app/navigation/navigation-activity/build.gradle.kts
@@ -1,10 +1,5 @@
 plugins {
-  id("hedvig.android.library")
   id("hedvig.android.ktlint")
+  id("hedvig.kotlin.library")
   alias(libs.plugins.squareSortDependencies)
-}
-
-dependencies {
-  implementation(projects.coreCommonAndroidPublic)
-  implementation(projects.coreResources)
 }
